nuking DB to restart. Got user info to appear on profile page.
