+++ imagersite/imagersite/views.py
@@ -28,33 +28,3 @@ def logout_view(request):
 def register_view(request):
     """View for registration."""
     return render(request, 'imagersite/register.html', context={})
-
-
-def profile_view(request):
-    """View for profile."""
-    if request.user.username == '':
-        return render(request, 'imagersite/profile.html', context={"logged_in": False})
-    else:
-        user_object = User.objects.get(username=request.user.username)
-        profile_object = Profile.objects.get(user=user_object)
-        profile_context_dict = {
-            "logged_in": True,
-            "user": user_object,
-            "profile": profile_object
-        }
-        return render(request, 'imagersite/profile.html', context=profile_context_dict)
-
-
-def alt_profile_view(request):
-    """View for non-user profile."""
-    try:
-        user_object = User.objects.get(username=request.path[1:])
-    except User.DoesNotExist:
-        return render(request, 'imagersite/altprofile.html', context={"user_exists": False})
-    profile_object = Profile.objects.get(user=user_object)
-    alt_profile_context_dict = {
-        "user_exists": True,
-        "user": user_object,
-        "profile": profile_object
-    }
-    return render(request, 'imagersite/altprofile.html', context=alt_profile_context_dict)
